rocket: update to be compatible with 0.5.1
The current GitHub and crates.io versions of rocket_anyhow are
incompatible with Rocket 0.5.1 and modern Rust due to having a
dependency tree littered with crates which were dependent on old
nightly features.

This patch updates the dependencies and makes a minor source change
to be compatible with Rocket's current API. Upon merging of this
patch, the crate on crates.io should be updated to allow other
crates pushed to crates.io to utilize these changes.

Signed-off-by: Amy Parker <amy@amyip.net>